requirements.txt
discord.py
requests
python-dotenv
sqlite3
